+++ datatest/src/commonMain/kotlin/at/asitplus/test/Intermediate.iterable.kt
@@ -12,6 +12,7 @@ import de.infix.testBalloon.framework.core.TestSuiteScope
  * @param compact If true, only a single test element is created and the class name of the data parameter is used as test name
  * @param maxLength maximum length of test element name (not display name)
  * @param displayNameMaxLength maximum length of test element **display name**
+ * @param prefix an optional prefix to add to the test name
  * @param testConfig Optional test configuration
  */
 fun <Data> TestSuiteScope.withData(
@@ -19,11 +20,13 @@ fun <Data> TestSuiteScope.withData(
     compact: Boolean = DataTest.compactByDefault,
     maxLength: Int = DataTest.defaultTestNameMaxLength,
     displayNameMaxLength: Int = DataTest.defaultDisplayNameMaxLength,
+    prefix: String = "",
     testConfig: TestConfig = TestConfig,
 ) = ConfiguredDataTestScope<Data>(
     compact,
     maxLength,
     displayNameMaxLength = displayNameMaxLength,
+    prefix = prefix,
     this,
     data.asSequence().map { it.toPrettyString() to it },
     testConfig
@@ -39,6 +42,7 @@ fun <Data> TestSuiteScope.withData(
  * @param compact If true, only a single test element is created and the class name of the data parameter is used as test name
  * @param maxLength maximum length of test element name (not display name)
  * @param displayNameMaxLength maximum length of test element **display name**
+ * @param prefix an optional prefix to add to the test name
  * @param testConfig Optional test configuration
  */
 fun <Data> TestSuiteScope.withData(
@@ -47,11 +51,13 @@ fun <Data> TestSuiteScope.withData(
     compact: Boolean = DataTest.compactByDefault,
     maxLength: Int = DataTest.defaultTestNameMaxLength,
     displayNameMaxLength: Int = DataTest.defaultDisplayNameMaxLength,
+    prefix: String = "",
     testConfig: TestConfig = TestConfig,
 ) = ConfiguredDataTestScope<Data>(
     compact,
     maxLength,
     displayNameMaxLength = displayNameMaxLength,
+    prefix = prefix,
     this,
     data.asSequence().map { nameFn(it) to it },
     testConfig
@@ -62,10 +68,11 @@ fun <Data> TestSuiteScope.withData(
  * Creates a test suite for each item in the provided iterable data.
  *
  * @param data The iterable collection of test data
- * @param testConfig Optional test configuration
  * @param compact If true, only a single test element is created and the class name of the data parameter is used as test name
  * @param maxLength maximum length of test element name (not display name)
  * @param displayNameMaxLength maximum length of test element **display name**
+ * @param testConfig Optional test configuration
+ * @param prefix an optional prefix to add to the test name
  * @param action Test suite configuration action for each data item
  */
 fun <Data> TestSuiteScope.withDataSuites(
@@ -73,6 +80,7 @@ fun <Data> TestSuiteScope.withDataSuites(
     compact: Boolean = DataTest.compactByDefault,
     maxLength: Int = DataTest.defaultTestNameMaxLength,
     displayNameMaxLength: Int = DataTest.defaultDisplayNameMaxLength,
+    prefix: String = "",
     testConfig: TestConfig = TestConfig,
     action: TestSuiteScope.(Data) -> Unit
 ) = withDataSuitesInternal(
@@ -80,6 +88,7 @@ fun <Data> TestSuiteScope.withDataSuites(
     compact,
     maxLength,
     displayNameMaxLength = displayNameMaxLength,
+    prefix = prefix,
     testConfig,
     action
 )
@@ -90,10 +99,11 @@ fun <Data> TestSuiteScope.withDataSuites(
  *
  * @param nameFn Function to generate suite name from data
  * @param data The iterable collection of test data
- * @param testConfig Optional test configuration
  * @param compact If true, only a single test element is created and the class name of the data parameter is used as test name
  * @param maxLength maximum length of test element name (not display name)
  * @param displayNameMaxLength maximum length of test element **display name**
+ * @param prefix an optional prefix to add to the test name
+ * @param testConfig Optional test configuration
  * @param action Test suite configuration action for each data item
  */
 fun <Data> TestSuiteScope.withDataSuites(
@@ -102,6 +112,7 @@ fun <Data> TestSuiteScope.withDataSuites(
     compact: Boolean = DataTest.compactByDefault,
     maxLength: Int = DataTest.defaultTestNameMaxLength,
     displayNameMaxLength: Int = DataTest.defaultDisplayNameMaxLength,
+    prefix: String = "",
     testConfig: TestConfig = TestConfig,
     action: TestSuiteScope.(Data) -> Unit
 ) = withDataSuitesInternal(
@@ -109,6 +120,7 @@ fun <Data> TestSuiteScope.withDataSuites(
     compact,
     maxLength,
     displayNameMaxLength = displayNameMaxLength,
+    prefix = prefix,
     testConfig,
     action
 )