+++ datatest/src/commonMain/kotlin/at/asitplus/test/Intermediate.map.kt
@@ -13,6 +13,7 @@ import de.infix.testBalloon.framework.core.TestSuiteScope
  * @param compact If true, only a single test element is created and the class name of the data parameter is used as test name
  * @param maxLength maximum length of test element name (not display name)
  * @param displayNameMaxLength maximum length of test element **display name**
+ * @param prefix an optional prefix to add to the test name
  * @param testConfig Optional test configuration
  */
 fun <Data> TestSuiteScope.withData(
@@ -20,11 +21,13 @@ fun <Data> TestSuiteScope.withData(
     compact: Boolean = DataTest.compactByDefault,
     maxLength: Int = DataTest.defaultTestNameMaxLength,
     displayNameMaxLength: Int = DataTest.defaultDisplayNameMaxLength,
+    prefix: String = "",
     testConfig: TestConfig = TestConfig,
 ) = ConfiguredDataTestScope<Data>(
     compact,
     maxLength,
     displayNameMaxLength = displayNameMaxLength,
+    prefix = prefix,
     this,
     map.asSequence().map { (k, v) -> k to v },
     testConfig
@@ -39,6 +42,7 @@ fun <Data> TestSuiteScope.withData(
  * @param compact If true, only a single test element is created and the class name of the data parameter is used as test name
  * @param maxLength maximum length of test element name (not display name)
  * @param displayNameMaxLength maximum length of test element **display name**
+ * @param prefix an optional prefix to add to the test name
  * @param action Test suite configuration action for each map value
  */
 fun <Data> TestSuiteScope.withDataSuites(
@@ -46,6 +50,7 @@ fun <Data> TestSuiteScope.withDataSuites(
     compact: Boolean = DataTest.compactByDefault,
     maxLength: Int = DataTest.defaultTestNameMaxLength,
     displayNameMaxLength: Int = DataTest.defaultDisplayNameMaxLength,
+    prefix: String = "",
     testConfig: TestConfig = TestConfig,
     action: TestSuiteScope.(Data) -> Unit
 ) = withDataSuitesInternal(
@@ -53,6 +58,7 @@ fun <Data> TestSuiteScope.withDataSuites(
     compact,
     maxLength,
     displayNameMaxLength = displayNameMaxLength,
+    prefix = prefix,
     testConfig,
     action
 )