+++ datatest/src/commonMain/kotlin/at/asitplus/test/Intermediate.vararg.kt
@@ -12,6 +12,7 @@ import de.infix.testBalloon.framework.core.TestSuiteScope
  * @param compact If true, only a single test element is created and the class name of the data parameter is used as test name
  * @param maxLength maximum length of test element name (not display name)
  * @param displayNameMaxLength maximum length of test element **display name**
+ * @param prefix an optional prefix to add to the test name
  * @param testConfig Optional test configuration
  */
 @Suppress("INVISIBLE_MEMBER", "INVISIBLE_REFERENCE")
@@ -21,11 +22,13 @@ fun <Data> TestSuiteScope.withData(
     compact: Boolean = DataTest.compactByDefault,
     maxLength: Int = DataTest.defaultTestNameMaxLength,
     displayNameMaxLength: Int = DataTest.defaultDisplayNameMaxLength,
+    prefix: String = "",
     testConfig: TestConfig = TestConfig
 ) = ConfiguredDataTestScope<Data>(
     compact,
     maxLength,
     displayNameMaxLength = displayNameMaxLength,
+    prefix = prefix,
     this,
     parameters.asSequence().map { it.toPrettyString() to it },
     testConfig
@@ -41,6 +44,7 @@ fun <Data> TestSuiteScope.withData(
  * @param compact If true, only a single test element is created and the class name of the data parameter is used as test name
  * @param maxLength maximum length of test element name (not display name)
  * @param displayNameMaxLength maximum length of test element **display name**
+ * @param prefix an optional prefix to add to the test name
  * @param testConfig Optional test configuration
  */
 @Suppress("INVISIBLE_MEMBER", "INVISIBLE_REFERENCE")
@@ -51,11 +55,13 @@ fun <Data> TestSuiteScope.withData(
     compact: Boolean = DataTest.compactByDefault,
     maxLength: Int = DataTest.defaultTestNameMaxLength,
     displayNameMaxLength: Int = DataTest.defaultDisplayNameMaxLength,
+    prefix: String = "",
     testConfig: TestConfig = TestConfig
 ) = ConfiguredDataTestScope<Data>(
     compact,
     maxLength,
     displayNameMaxLength = displayNameMaxLength,
+    prefix = prefix,
     this,
     parameters.asSequence().map { nameFn(it) to it },
     testConfig
@@ -66,10 +72,11 @@ fun <Data> TestSuiteScope.withData(
  * Creates a test suite for each provided data parameter.
  *
  * @param parameters The data parameters to create suites for
- * @param testConfig Optional test configuration
  * @param compact If true, only a single test element is created and the class name of the data parameter is used as test name
  * @param maxLength maximum length of test element name (not display name)
  * @param displayNameMaxLength maximum length of test element **display name**
+ * @param prefix an optional prefix to add to the test name
+ * @param testConfig Optional test configuration
  * @param action Test suite configuration action for each parameter
  */
 @Suppress("INVISIBLE_MEMBER", "INVISIBLE_REFERENCE")
@@ -79,6 +86,7 @@ fun <Data> TestSuiteScope.withDataSuites(
     compact: Boolean = DataTest.compactByDefault,
     maxLength: Int = DataTest.defaultTestNameMaxLength,
     displayNameMaxLength: Int = DataTest.defaultDisplayNameMaxLength,
+    prefix: String = "",
     testConfig: TestConfig = TestConfig,
     action: TestSuiteScope.(Data) -> Unit
 ) = withDataSuitesInternal(
@@ -86,6 +94,7 @@ fun <Data> TestSuiteScope.withDataSuites(
     compact,
     maxLength,
     displayNameMaxLength = displayNameMaxLength,
+    prefix = prefix,
     testConfig,
     action
 )
@@ -97,10 +106,11 @@ fun <Data> TestSuiteScope.withDataSuites(
  *
  * @param nameFn Function to generate suite name from data
  * @param parameters The data parameters to create suites for
- * @param testConfig Optional test configuration
  * @param compact If true, only a single test element is created and the class name of the data parameter is used as test name
  * @param maxLength maximum length of test element name (not display name)
  * @param displayNameMaxLength maximum length of test element **display name**
+ * @param prefix an optional prefix to add to the test name
+ * @param testConfig Optional test configuration
  * @param action Test suite configuration action for each parameter
  */
 @Suppress("INVISIBLE_MEMBER", "INVISIBLE_REFERENCE")
@@ -111,6 +121,7 @@ fun <Data> TestSuiteScope.withDataSuites(
     compact: Boolean = DataTest.compactByDefault,
     maxLength: Int = DataTest.defaultTestNameMaxLength,
     displayNameMaxLength: Int = DataTest.defaultDisplayNameMaxLength,
+    prefix: String = "",
     testConfig: TestConfig = TestConfig,
     action: TestSuiteScope.(Data) -> Unit
 ) = withDataSuitesInternal(
@@ -118,6 +129,7 @@ fun <Data> TestSuiteScope.withDataSuites(
     compact,
     maxLength,
     displayNameMaxLength = displayNameMaxLength,
+    prefix = prefix,
     testConfig,
     action
 )