+++ datatest/src/commonMain/kotlin/at/asitplus/test/Terminal.map.kt
@@ -14,6 +14,7 @@ import de.infix.testBalloon.framework.core.TestSuiteScope
  * @param compact If true, only a single test element is created and the class name of the data parameter is used as test name
  * @param maxLength maximum length of test element name (not display name)
  * @param displayNameMaxLength maximum length of test element **display name**
+ * @param prefix an optional prefix to add to the test name
  * @param testConfig Optional test configuration
  * @param action Test action to execute for each map value
  */
@@ -22,6 +23,7 @@ fun <Data> TestSuiteScope.withData(
     compact: Boolean = DataTest.compactByDefault,
     maxLength: Int = DataTest.defaultTestNameMaxLength,
     displayNameMaxLength: Int = DataTest.defaultDisplayNameMaxLength,
+    prefix: String = "",
     testConfig: TestConfig = TestConfig,
     action: suspend Test.ExecutionScope.(Data) -> Unit
 ) = withDataInternal(
@@ -30,6 +32,7 @@ fun <Data> TestSuiteScope.withData(
     compact,
     maxLength,
     displayNameMaxLength,
+    prefix,
     action
 )