+++ datatest/src/commonMain/kotlin/at/asitplus/test/Terminal.vararg.kt
@@ -2,7 +2,6 @@ package at.asitplus.testballoon
 
 import de.infix.testBalloon.framework.core.Test
 import de.infix.testBalloon.framework.core.TestConfig
-import de.infix.testBalloon.framework.core.TestSuite
 import de.infix.testBalloon.framework.core.TestSuiteScope
 
 /**
@@ -12,6 +11,7 @@ import de.infix.testBalloon.framework.core.TestSuiteScope
  * @param compact If true, only a single test element is created and the class name of the data parameter is used as test name
  * @param maxLength maximum length of test element name (not display name)
  * @param displayNameMaxLength maximum length of test element **display name**
+ * @param prefix an optional prefix to add to the test name
  * @param testConfig Optional test configuration
  * @param action Test action to execute for each parameter
  */
@@ -22,6 +22,7 @@ fun <Data> TestSuiteScope.withData(
     compact: Boolean = DataTest.compactByDefault,
     maxLength: Int = DataTest.defaultTestNameMaxLength,
     displayNameMaxLength: Int = DataTest.defaultDisplayNameMaxLength,
+    prefix: String = "",
     testConfig: TestConfig = TestConfig,
     action: suspend Test.ExecutionScope.(Data) -> Unit
 ) = withDataInternal(
@@ -30,6 +31,7 @@ fun <Data> TestSuiteScope.withData(
     compact,
     maxLength,
     displayNameMaxLength,
+    prefix,
     action
 )
 
@@ -38,10 +40,11 @@ fun <Data> TestSuiteScope.withData(
  * Uses provided function to generate test names.
  *
  * @param nameFn Function to generate test name from data
- * @param arguments The data parameters to test with
+ * @param parameters The data parameters to test with
  * @param compact If true, only a single test element is created and the class name of the data parameter is used as test name
  * @param maxLength maximum length of test element name (not display name)
  * @param displayNameMaxLength maximum length of test element **display name**
+ * @param prefix an optional prefix to add to the test name
  * @param testConfig Optional test configuration
  * @param action Test action to execute for each parameter
  */
@@ -53,6 +56,7 @@ fun <Data> TestSuiteScope.withData(
     compact: Boolean = DataTest.compactByDefault,
     maxLength: Int = DataTest.defaultTestNameMaxLength,
     displayNameMaxLength: Int = DataTest.defaultDisplayNameMaxLength,
+    prefix: String = "",
     testConfig: TestConfig = TestConfig,
     action: suspend Test.ExecutionScope.(Data) -> Unit
 ) = withDataInternal(
@@ -61,5 +65,6 @@ fun <Data> TestSuiteScope.withData(
     compact,
     maxLength,
     displayNameMaxLength,
+    prefix,
     action
 )